requirements.txt
selenium
beautifulsoup4
pillow
numpy
opencv-python
pytesseract
pyautogui
PyQt5
